+++ src/background/name-search.ts
@@ -29,6 +29,10 @@ export async function nameSearch({
   // Record the position of existing entries for grouping purposes
   const existingItems = new Map<string, number>();
 
+  // Record which entries we have already seen so we don't try to merge the same
+  // entries when matching on variants
+  const have = new Set<number>();
+
   let currentString = input;
 
   while (currentString.length > 0) {
@@ -66,6 +70,8 @@ export async function nameSearch({
         return null;
       }
 
+      // Filter out entries we already have
+      names = names.filter((name) => !have.has(name.id));
       if (!names.length) {
         continue;
       }
@@ -73,6 +79,8 @@ export async function nameSearch({
       result.matchLen = Math.max(result.matchLen, currentInputLength);
 
       for (const name of names) {
+        have.add(name.id);
+
         // We group together entries where the kana readings and translation
         // details are all equal.
         const nameContents = getNameEntryHash(name);